edited text for readability and compatibility with the style guide
This commit corrects facts (e.g. Terraform is a tool, not a language) and clarifies that the purpose of the instruction set is to create an NGINX server using Docker, not AWS. Furthermore, context was given for each of the three major sections and an additional step was included to launch Docker. Finally, the text was edited for active language, present tense and general appropriateness.